+++ src/scripts/hooks/index.ts
@@ -3,4 +3,5 @@ export { useEventListener } from './modules/useEventListener'
 export { useMountedRef } from './modules/useMountedRef'
 export { useSnackbarContext } from './modules/useSnackbarContext'
 export { usePrevious } from './modules/usePrevious'
-export { useFetchBoardAndList } from './modules/useFetchBoardAndList'
+export { useFetchBoard } from './modules/useFetchBoard'
+export { useFetchList } from './modules/useFetchList'
